requirements.txt
flask
tensorflow
keras
pillow
numpy
